refactor: split forwarder and add main entry
